Moved desired capabilities to AbstractWebDriverRequest
